+++ src/App/Views/User/Password/style.js
@@ -1,4 +1,4 @@
-import { cardTitle } from '../../Assets/Jss/material-dashboard-pro-react';
+import { cardTitle } from '../../../Assets/Jss/material-dashboard-pro-react';
 
 const userProfileStyles = {
   cardTitle,
